The DTO layer has been completed. Access modifiers have been modified.

+++ src/main/java/com/github/kutik93/service/ProjectRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface ProjectRepository {
     List<Project> findAll();
-    Optional<Project> findById(Integer id);
+    Optional<Project> findById(Long id);
     Project save(Project entity);
 }
